pyproject.toml
[project]
name = "llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kazuya-yonekura", email = "kazuya.yonekura@sencorp.co.jp" }
]
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.0",
    "llama-cloud-services>=0.6.1",
    "llama-index>=0.12.15",
    "llama-index-core>=0.12.15",
    "llama-index-embeddings-ollama>=0.5.0",
    "llama-index-readers-file>=0.4.4",
    "llmsherpa>=0.1.4",
    "ollama>=0.4.7",
    "python-dotenv>=1.0.1",
]

[project.scripts]
llamaindex = "llamaindex:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
